pyproject.toml
[tool.poetry]
name = "nonebot-plugin-blockwords"
version = "0.1.2"
description = "nonebot聊天屏蔽词插件"
authors = ["MelodyKnit <2711402357@qq.com>"]
license = "MIT"
keywords = [
    "bot",
    "nonebot2",
    "nonebot-plugin",
    "blockwords",
]
readme = "README.md"
packages = [{ include = "nonebot_plugin_blockwords" }]

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.1.1"
jieba = "^0.42.1"

[tool.poetry.group.dev.dependencies]
isort = "^5.10.1"
black = "^23.1.0"
pre-commit = "^3.0.0"

[tool.poetry.group.test.dependencies]
nonebug = "^0.3.5"
pytest-asyncio = "^0.21.1"
nonebot-adapter-console = "^0.4.0"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
priority = "default"

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 88
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
src_paths = ["nonebot_plugin_blockwords"]
extra_standard_library = ["typing_extensions"]

[tool.pyright]
pythonVersion = "3.8"
pythonPlatform = "All"
executionEnvironments = [
    { root = "./" },
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
